Add test file and restructure of project
