+++ AlphanumericStringSort/src/app.ts
@@ -1,0 +1,5 @@
+import { sortString } from "./assets/functions";
+
+const inputString = "aB3c9!2x@C%$%SFSDdsf";
+const sortedString = sortString(inputString);
+console.log(sortedString);
