+++ AlphanumericStringSort/src/interfaces/sortedArrays.ts
@@ -1,0 +1,5 @@
+export interface SortedArrays {
+    letters: string[];
+    numbers: string[];
+    specialCharacters: string[];
+}
